feat: dogstatsd and defaulting
Implemented defaulting for webhooks, set up the local dev environment to
test for it, and set up a basic dogstatsd output inline with the
scraper.  Next steps will be to remove the inline nature of the scraper
and replace it with a generator with configurable concurrency for both
the input and outputs.